Environment

Description

This module contains all the auxiliary logic necessary to represent the internal state of the interpreter.

Environment

data Environment

context :: Environment -> Context

defaultEnv :: Environment

Default environment for the interpreter.

emptyContext :: Context

Empty context without any bindings

Reading the environment

getVerbose :: Environment -> Bool

Get current settings

getColor :: Environment -> Bool

Get current settings

getSki :: Environment -> Bool

Get current settings

getTypes :: Environment -> Bool

Get current settings

getExpressionName :: Environment -> Exp -> Maybe String

Given an expression, returns its name if it is bounded to any.

getTopo :: Environment -> Bool

Get current settings

Modifying the environment

addBind :: Environment -> String -> Exp -> Environment

Adds a name binding to the environment

changeColor :: Environment -> Bool -> Environment

Sets the color configuration on/off

changeVerbose :: Environment -> Bool -> Environment

Sets the verbose configuration on/off.

changeSkioutput :: Environment -> Bool -> Environment

Sets the ski output configuration on/off

changeTypes :: Environment -> Bool -> Environment

Sets the types output configuration on/off

changeTopo :: Environment -> Bool -> Environment

Filenames and Modulenames

type Filename = String

A filename is a string containing the directory path and the real name of the file.

type Modulename = String

A modulename is a string naming a module.