Format

Description

This module controls the format of the text and expressions printed by the interpreter. Uses ANSI escape sequences to color the terminal and mark text as bold or italics. It also stores the texts showed by the interpreter.

Formatting

formatFormula :: String

Sequence of characters that signals the format of a formula to the terminal.

formatIntro :: String

Sequence of characters that signals the format of a formula to the terminal.

formatLoading :: String

Sequence of characters that signals the format of a formula to the terminal.

formatPrompt :: String

Sequence of characters that signals the format of a formula to the terminal.

formatName :: String

Sequence of characters that signals the format of a formula to the terminal.

formatSubs1 :: String

formatSubs2 :: String

formatType :: String

decolor :: String -> String

Removes all color from a string

end :: String

Sequence of characters that cleans all the format.

Interpreter texts

promptText :: String

Prompt line. It is shown when the interpreter asks the user to introduce a new command.

helpText :: String

Help line. It is shown when the user uses the :help command.

initialText :: String

Initial text on the interpreter. It is shown at startup.

versionText :: String

Version complete text

restartText :: String

errorNonTypeableText :: String

Non typeable expression error message.

errorTypeConstructors :: String

Type constructors on untyped lambda calculus error message.

errorUndefinedText :: String

Undefined expression error message.

errorUnknownCommand :: String

Unknown command error message