Interpreter

Description

This module contains auxiliary logic, types and representations of the internal state of the interpreter.

Interpreter actions

data InterpreterAction

Interpreter action. It can be a language action (binding and evaluation) or an interpreter specific one, such as "quit".

Constructors

Interpret Action
Language action

EmptyLine
Empty line, it will be ignored

Error
Error on the interpreter

Quit
Close the interpreter

Restart
Restarts the environment

Load String
Load the given file

SetVerbose Bool
Changes verbosity

SetColor Bool
Changes colors

SetSki Bool
Changes ski output

SetTypes Bool
Changes type configuration

SetTopo Bool

Help
Shows help

interpreteractionParser :: Parser InterpreterAction

Parses an interpreter action.

act :: Action -> State Environment [String]

Executes a language action. Given a context and an action, returns the new context after the action and a text output.

multipleAct :: [Action] -> State Environment [String]

Executes multiple actions. Given a context and a set of actions, returns the new context after the sequence of actions and a text output.

data Action

Language action. The language has a number of possible valid statements; all on the following possible forms.

Constructors

Bind (String, NamedLambda)
bind a name to an expression

EvalBind (String, NamedLambda)
bind a name to an expression and simplify it

Execute NamedLambda
execute an expression

Comment
comment

actionParser :: Parser Action

Parses a language action.