Main

Documentation

main :: IO ()

Runs the interpreter with default settings and an empty context.

interpreterLoop :: Environment -> InputT IO ()

Interpreter awaiting for an instruction.

setOption :: Environment -> Bool -> (Environment -> Bool -> Environment) -> String -> InputT IO ()

Sets the given option on/off.

outputActions :: Environment -> [String] -> InputT IO ()

Outputs results from actions. Given a list of options and outputs, formats and prints them in console.

loadFile :: String -> IO (Maybe [Action])

Loads the given filename and returns the complete list of actions. Returns Nothing if there is an error reading or parsing the file.

executeFile :: Filename -> IO ()

Executes the commands inside a file. A .mkr file can contain a sequence of expressions and variable bindings, and it is interpreted sequentially.

readFileDependencies :: Filename -> IO [Modulename]

Reads module dependencies

readAllModuleDeps :: [Modulename] -> IO [Modulename]

Reads all the dependencies from a module list

readAllModuleDepsRecursively :: [Modulename] -> IO [Modulename]

Read module dependencies recursively

findFilename :: Modulename -> IO Filename

Given a module name, returns the filename associated with it

data MainFlags

Flags datatype

Constructors

MainFlags

Fields

flagExec :: String

flagVersion :: Bool