Ski

Description

This module implements a representation of the SKI subset of the calculus of combinators. It provides a lambda abstraction algorithm writing lambda expressions as combinators.

Documentation

data Ski

A SKI combinator expression

Constructors

S

K

I

Comb Ski Ski

Cte String

Spair

Spi1

Spi2

Sinl

Sinr

Scase

Sunit

Sabort

Sabsurd

skiabs :: NamedLambda -> Ski

SKI abstraction of a named lambda term. From a lambda expression creates a SKI equivalent expression. The following algorithm is a version of the algorithm 9.10 on the Hindley-Seldin book.