Stlc.Types

Documentation

data Type

A type template is a free type variable or an arrow between two types; that is, the function type.

Constructors

Tvar Variable

Arrow Type Type

Times Type Type

Union Type Type

Unitty

Bottom

typeinference :: Exp -> Maybe Type

Type inference of a lambda expression.

typeinfer

Arguments

:: [Variable]
List of fresh variables

-> Context
Type context

-> Exp
Lambda expression whose type has to be inferred

-> Type
Constraint

-> Maybe Substitution

Type inference algorithm. Infers a type from a given context and expression with a set of constraints represented by a unifier type. The result type must be unifiable with this given type.

unify :: Type -> Type -> Maybe Substitution

Unifies two types with their most general unifier. Returns the substitution that transforms any of the types into the unifier.

applyctx :: Substitution -> Context -> Context

Apply a substitution to all the types on a type context.

emptyctx :: Context

The empty context.

incrementindices :: Context -> Context

Increments all the indices of a given context. It is useful for adapting the context to a new scope.

variables :: [Variable]

Infinite list of variables.

normalizeTemplate :: Map Integer Integer -> Integer -> Type -> (Map Integer Integer, Integer)

Substitutes a set of type variables on a type template for the smaller possible ones.

applynormalization :: Map Integer Integer -> Type -> Type

Applies a set of variable substitutions to a type to normalize it.

newtype Top

Constructors

Top Type

type Context = Map Integer Type

A type context is a map from deBruijn indices to types. Given any lambda variable as a deBruijn index, it returns its type.

type Variable = Integer

A type variable is an integer.

type Substitution = Type -> Type

A type substitution is a function that can be applied to any type to get a new one.