Main

Documentation

main :: IO ()

Runs the interpreter with default settings and an empty context.

interpreterLoop :: Environment -> InputT IO ()

Interpreter awaiting for an instruction.

executeAction :: Environment -> InterpreterAction -> InputT IO (Maybe Environment)

Executes the parsed action, every action may affect the context in a way, and returns the control to the interpreter.

outputActions :: Environment -> [String] -> InputT IO ()

Outputs results from actions. Given a list of options and outputs, formats and prints them in console.

loadFile :: Filename -> IO (Maybe [Action])

Loads the given filename and returns the complete list of actions. Returns Nothing if there is an error reading or parsing the file.

executeFile :: Filename -> IO ()

Executes the commands inside a file. A .mkr file can contain a sequence of expressions and variable bindings, and it is interpreted sequentially.

readFileDependencies :: Filename -> IO [Modulename]

Reads module dependencies

readAllModuleDeps :: [Modulename] -> IO (Maybe [Modulename])

Reads all the dependencies from a module list. Returns an error if a dependency cannot be found

readAllModuleDepsRecursively :: [Modulename] -> IO (Maybe [Modulename])

Read module dependencies recursively. Returns an error if a dependency cannot be found

findFilename :: Modulename -> IO (Maybe Filename)

Given a module name, returns the filename associated with it

data MainFlags

Flags datatype

Constructors

MainFlags

Fields

flagExec :: String

flagVersion :: Bool

flagLibs :: Bool